26 assignments page (#34)
* implementing the assignment page route

* Update header.component.scss

* feat: Add overlay to close pop-up and improve pop-up animations in AssignmentsPageComponent

Added overlay to close the popup when clicking outside of it.
Implemented smooth entry animation for the popup.
Updated closePopup method to manage closing.
CSS adjustments to support overlay and center the pop-up.
Added support for opening popup based on slug via query parameters.

* fix: Ensure pop-up functionality works seamlessly when navigating from home to ContentPageComponent

Adjust ngOnInit to automatically open the corresponding pop-up when browsing via query parameters.
Updates to the openPopupBySlug method to correctly find and display the selected activity.
Guaranteed that the "Learn More" navigation functionality opens the specific pop-up when arriving on the page.